refactor qaqc
